+++ tutor/commands/dev.py
@@ -1,4 +1,3 @@
-import os
 from typing import List
 
 import click
@@ -6,39 +5,40 @@ import click
 from .. import config as tutor_config
 from .. import env as tutor_env
 from .. import fmt
-from ..types import Config
-from .. import utils
+from ..types import Config, get_typed
 from . import compose
-from .context import Context
 
 
-def docker_compose(root: str, config: Config, *command: str) -> int:
-    """
-    Run docker-compose with dev arguments.
-    """
-    args = []
-    for folder in ["local", "dev"]:
-        # Add docker-compose.yml and docker-compose.override.yml (if it exists)
-        # from "local" and "dev" folders (but not docker-compose.prod.yml)
-        args += [
-            "-f",
-            tutor_env.pathjoin(root, folder, "docker-compose.yml"),
+class DevJobRunner(compose.ComposeJobRunner):
+    def __init__(self, root: str, config: Config):
+        """
+        Load docker-compose files from dev/ and local/
+        """
+        super().__init__(root, config)
+        self.project_name = get_typed(self.config, "DEV_PROJECT_NAME", str)
+        self.docker_compose_files += [
+            tutor_env.pathjoin(self.root, "local", "docker-compose.yml"),
+            tutor_env.pathjoin(self.root, "dev", "docker-compose.yml"),
+            tutor_env.pathjoin(self.root, "local", "docker-compose.override.yml"),
+            tutor_env.pathjoin(self.root, "dev", "docker-compose.override.yml"),
         ]
-        override_path = tutor_env.pathjoin(root, folder, "docker-compose.override.yml")
-        if os.path.exists(override_path):
-            args += ["-f", override_path]
-    return utils.docker_compose(
-        *args,
-        "--project-name",
-        str(config["DEV_PROJECT_NAME"]),
-        *command,
-    )
+        self.docker_compose_job_files += [
+            tutor_env.pathjoin(self.root, "local", "docker-compose.jobs.yml"),
+            tutor_env.pathjoin(self.root, "dev", "docker-compose.jobs.yml"),
+            tutor_env.pathjoin(self.root, "local", "docker-compose.jobs.override.yml"),
+            tutor_env.pathjoin(self.root, "dev", "docker-compose.jobs.override.yml"),
+        ]
+
+
+class DevContext(compose.BaseComposeContext):
+    def job_runner(self, config: Config) -> DevJobRunner:
+        return DevJobRunner(self.root, config)
 
 
 @click.group(help="Run Open edX locally with development settings")
-@click.pass_obj
-def dev(context: Context) -> None:
-    context.docker_compose_func = docker_compose
+@click.pass_context
+def dev(context: click.Context) -> None:
+    context.obj = DevContext(context.obj.root)
 
 
 @click.command(