+++ tutor/commands/local.py
@@ -1,4 +1,3 @@
-import os
 from time import sleep
 
 import click
@@ -6,38 +5,40 @@ import click
 from .. import config as tutor_config
 from .. import env as tutor_env
 from .. import fmt
-from ..types import get_typed, Config
+from ..types import Config, get_typed
 from .. import utils
 from .. import exceptions
 from . import compose
 from .config import save as config_save_command
-from .context import Context
 
 
-def docker_compose(root: str, config: Config, *command: str) -> int:
-    """
-    Run docker-compose with local and production yml files.
-    """
-    args = []
-    override_path = tutor_env.pathjoin(root, "local", "docker-compose.override.yml")
-    if os.path.exists(override_path):
-        args += ["-f", override_path]
-    return utils.docker_compose(
-        "-f",
-        tutor_env.pathjoin(root, "local", "docker-compose.yml"),
-        "-f",
-        tutor_env.pathjoin(root, "local", "docker-compose.prod.yml"),
-        *args,
-        "--project-name",
-        get_typed(config, "LOCAL_PROJECT_NAME", str),
-        *command
-    )
+class LocalJobRunner(compose.ComposeJobRunner):
+    def __init__(self, root: str, config: Config):
+        """
+        Load docker-compose files from local/.
+        """
+        super().__init__(root, config)
+        self.project_name = get_typed(self.config, "LOCAL_PROJECT_NAME", str)
+        self.docker_compose_files += [
+            tutor_env.pathjoin(self.root, "local", "docker-compose.yml"),
+            tutor_env.pathjoin(self.root, "local", "docker-compose.prod.yml"),
+            tutor_env.pathjoin(self.root, "local", "docker-compose.override.yml"),
+        ]
+        self.docker_compose_job_files += [
+            tutor_env.pathjoin(self.root, "local", "docker-compose.jobs.yml"),
+            tutor_env.pathjoin(self.root, "local", "docker-compose.jobs.override.yml"),
+        ]
+
+
+class LocalContext(compose.BaseComposeContext):
+    def job_runner(self, config: Config) -> LocalJobRunner:
+        return LocalJobRunner(self.root, config)
 
 
 @click.group(help="Run Open edX locally with docker-compose")
-@click.pass_obj
-def local(context: Context) -> None:
-    context.docker_compose_func = docker_compose
+@click.pass_context
+def local(context: click.Context) -> None:
+    context.obj = LocalContext(context.obj.root)
 
 
 @click.command(help="Configure and run Open edX from scratch")